Backend/requirements.txt
# --- Project Core & Framework ---
fastapi==0.129.0
uvicorn==0.40.0
fastapi-cli==0.0.21
fastapi-cloud-cli==0.12.0
python-multipart==0.0.22
starlette==0.52.1
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic-extra-types==2.11.0
pydantic_core==2.41.5

# --- AI, ML & Data Processing ---
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchaudio==2.6.0+cu124
numpy==2.4.2
pandas==2.3.3
matplotlib==3.10.8
altair==6.0.0
pyarrow==23.0.1
pillow==12.1.1
protobuf==6.33.5
sympy==1.13.1
mpmath==1.3.0
networkx==3.6.1

# --- Database (MongoDB Async) ---
motor==3.6.1
pymongo==4.9.2
dnspython==2.8.0

# --- API, Search & Web Tools ---
requests==2.32.5
google-search-results==2.4.2
duckduckgo_search==8.1.1
httpx==0.28.1
httptools==0.7.1
httpcore==1.0.9
urllib3==2.6.3
h11==0.16.0
websockets==16.0
beautifulsoup4==4.14.3
soupsieve==2.8.3

# --- Security, Auth & Environment ---
PyJWT==2.12.1
bcrypt==5.0.0
email-validator==2.3.0
python-dotenv==1.2.1

# --- UI, Dashboard & Reporting ---
streamlit==1.54.0
pydeck==0.9.1
fpdf==1.7.2
reportlab==4.4.10
rich==14.3.2
rich-toolkit==0.19.4

# --- System & Development Utilities ---
watchfiles==1.1.1
watchdog==6.0.0
click==8.3.1
typer==0.23.1
shellingham==1.5.4
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
MarkupSafe==3.0.3
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
attrs==25.4.0
anyio==4.12.1
sniffio==1.3.1
Pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2
blinker==1.9.0
cachetools==6.2.6
kiwisolver==1.4.9
pyparsing==3.3.2
contourpy==1.3.3
fonttools==4.61.1
six==1.17.0
packaging==26.0
tqdm==4.67.3
filelock==3.25.2

# --- Git & Monitoring ---
GitPython==3.1.46
gitdb==4.0.12
smmap==5.0.2
sentry-sdk==2.53.0
rignore==0.7.6

# --- Specific Additions ---
annotated-doc==0.0.4
annotated-types==0.7.0
fastar==0.8.0
narwhals==2.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
jiter==0.13.0